Quick search by holding Alt-key
Search for a title:     Hold down Alt-key and type a text. 
New search:      Release the Alt-key and repeat the process above. 

This is not a filtering tool, it's just a quick search tool for a game title, where a typed text pattern is at first compared with title staring-with text, then only when there is no match, the tool is searching again, this time for any text match. 
TODO: 
Someone may decide to add functionality for displaying a fancy looking search text pattern overlaying somewhere in the middle (or the top) of the screen to see what has just been typed (jumpToPattern string). New event handling Keys.onReleased/Alt-key would dismiss the overlay.